+++ test-data/projects/mono-repo/projects/a/src/a/core.clj
@@ -1,0 +1,4 @@
+(ns a.core)
+
+(defn function-a []
+  (println "output"))