+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- Fix: [Default clojure-lsp startup behaviour is changed since v2.0.327](https://github.com/BetterThanTomorrow/calva/issues/2084)
 - Fix: [clojure-lsp does not start in a workspace without project files](https://github.com/BetterThanTomorrow/calva/issues/2069)
 
 ## [2.0.332] - 2023-02-15