+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.333] - 2023-02-21
+
+- Fix: [Multiple LSP processes automatically running, broken LSP features in multi-root workspaces](https://github.com/BetterThanTomorrow/calva/issues/2065)
+- Fix: [Default clojure-lsp startup behaviour is changed since v2.0.327](https://github.com/BetterThanTomorrow/calva/issues/2084)
+- Fix: [clojure-lsp does not start in a workspace without project files](https://github.com/BetterThanTomorrow/calva/issues/2069)
+
 ## [2.0.332] - 2023-02-15
 
 - Partly fixes: [Cursor moves back to start of indented line when typing inside parens on a new line](https://github.com/BetterThanTomorrow/calva/issues/2071)